Move page layout helpers into seperate helper file
