+++ app/views/layouts/application.html.erb
@@ -26,14 +26,14 @@
     <meta property="og:image"        content="<%= page_image %>">
 
     <%#  Twitter Card - https://dev.twitter.com/cards/types/summary %>
-    <meta name="twitter:card"         content="summary">
-    <meta name="twitter:site"         content="@uofalibraries">
-    <meta name="twitter:url"          content="<%= request.base_url + request.fullpath %>">
-    <meta name="twitter:title"        content="<%= page_title %>">
-    <meta name="twitter:description"  content="<%= page_description %>">
-    <meta name="twitter:image"        content="<%= page_image %>">
-    <%= yield(:extra_social_meta) %>
+    <meta name="twitter:card"        content="summary">
+    <meta name="twitter:site"        content="@uofalibraries">
+    <meta name="twitter:url"         content="<%= request.base_url + request.fullpath %>">
+    <meta name="twitter:title"       content="<%= page_title %>">
+    <meta name="twitter:description" content="<%= page_description %>">
+    <meta name="twitter:image"       content="<%= page_image %>">
 
+    <%= yield(:extra_social_meta) %>
 
     <%= canonical_tag -%>
